fix: 시 단위로 정보 수정, CLGeocoder 호출 위치 수정(함수-> 뷰모델), LocationManager 수정
