+++ TarTanning/TarTanning/Core/Managers/LocationManager.swift
@@ -6,3 +6,48 @@
 //
 
 import Foundation
+import CoreLocation
+import Combine
+
+class LocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
+    private var locationManager = CLLocationManager()
+
+    @Published var latitude: Double = 0.0
+    @Published var longitude: Double = 0.0
+    @Published var locality: String = ""
+
+    private let geocoder = CLGeocoder()
+
+    override init() {
+        super.init()
+        locationManager.delegate = self
+        locationManager.desiredAccuracy = kCLLocationAccuracyBest
+        locationManager.requestWhenInUseAuthorization()
+    }
+
+    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
+        switch manager.authorizationStatus {
+        case .authorizedWhenInUse, .authorizedAlways:
+            locationManager.startUpdatingLocation()
+        case .notDetermined:
+            locationManager.requestWhenInUseAuthorization()
+        default:
+            break
+        }
+    }
+
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        guard let location = locations.last else { return }
+
+        geocoder.reverseGeocodeLocation(location) { placemarks, error in
+            guard let placemark = placemarks?.first, error == nil else { return }
+            let newLocality = placemark.locality ?? ""
+
+            if newLocality != self.locality {
+                self.latitude = location.coordinate.latitude
+                self.longitude = location.coordinate.longitude
+                self.locality = newLocality
+            }
+        }
+    }
+}
